requirements.txt
configparser==3.5.0
pathlib==1.0.1
PyGithub==1.39
PyJWT==1.6.4
